Move some test code around
