+++ src/test/scala/numscala/SpiralDataVisualTest.scala
@@ -1,0 +1,12 @@
+package numscala
+
+import numscala.SpiralData.calculateDataMatrix
+
+object SpiralDataVisualTest {
+
+    // Visual Test
+    def main(args: Array[String]): Unit = {
+        val answer = calculateDataMatrix()
+        println(answer.value.size)
+    }
+}
